% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_srcdyn.R
\name{calculate_dist}
\alias{calculate_dist}
\title{Calculate distance measures}
\usage{
calculate_dist(
  x,
  profile_id = profile_id,
  dates = dates,
  intensity = intensity,
  method = "euclidean"
)
}
\arguments{
\item{x}{a tibble in long format.}

\item{profile_id}{column name of the profile IDs (default is profile_id).}

\item{dates}{column name of the dates (default is dates).}

\item{intensity}{column name of the intensity (default is intensity).}

\item{method}{the distance measure to be used. This must be one of "euclidean",
"maximum", "manhattan", "canberra", "binary" or "minkowski" (default is "euclidean").}
}
\value{
Distance measure calculated on the scaled data.
}
\description{
This function takes a dataframe with profile IDs (numeric or character),
dates (date format) and intensity (numeric) and calculates distance measure
on the scaled data.
}
