% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_srcdyn.R
\name{compute_int_spr}
\alias{compute_int_spr}
\title{Compute intensity spread}
\usage{
compute_int_spr(x, profile_id = profile_id, intensity = intensity)
}
\arguments{
\item{x}{a tibble in long format.}

\item{profile_id}{column name of the profile IDs (default is profile_id).}

\item{intensity}{column name of the intensity (default is intensity).}
}
\value{
a tibble with profile IDs and their corresponding
intensity spread.
}
\description{
This function takes a dataframe with profile IDs (numeric or character) and
intensity (numeric) and computes intensity spread of time profiles as described
in Anliker et al. 2020.
}
\references{
Anliker, S., Loos, M., Comte, R., Ruff, M., Fenner, K.,
Singer, H., 2020. Assessing Emissions from Pharmaceutical Manufacturing
Based on Temporal High-Resolution Mass Spectrometry Data. Environ. Sci.
Technol. 54, 4110–4120. https://doi.org/10.1021/acs.est.9b07085
}
