% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_srcdyn.R
\name{detect_breaks}
\alias{detect_breaks}
\title{Detect breaks in time series}
\usage{
detect_breaks(x, profile_id = profile_id, dates = dates, intensity = intensity)
}
\arguments{
\item{x}{a tibble in long format.}

\item{profile_id}{column name of the profile IDs (default is profile_id).}

\item{dates}{column name of the dates (default is dates).}

\item{intensity}{column name of the intensity (default is intensity).}
}
\value{
A two-columns tibble with profile IDs and
the corresponding longest break detected.
}
\description{
This function takes a dataframe with profile IDs (numeric or character),
dates (date format) and intensity (numeric) and detects activity breaks in
time profiles.
}
\references{
Chonova, T., Honti, M., Loos, M., Ruppe, S., Langlois, I., Griesshaber, D.,
Fenner, K., Singer, H. Unveiling Industrial Contamination in the Rhine:
Insights from Data Mining of High-Frequency Measurements (in prep)
}
