% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_srcdyn.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot time profile}
\usage{
plot_profile(
  x,
  profile_id = profile_id,
  dates = dates,
  intensity = intensity,
  sel_prof
)
}
\arguments{
\item{x}{a tibble in long format.}

\item{profile_id}{column name of the profile IDs (default is profile_id).}

\item{dates}{column name of the dates (default is dates).}

\item{intensity}{column name of the intensity (default is intensity).}

\item{sel_prof}{the ID of the selected profile.}
}
\value{
A line plot of the selected time profile
}
\description{
This function takes a dataframe with profile IDs (numeric or character),
dates (date format) and intensity (numeric) and plots a selected time profile.
}
