% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_srcdyn.R
\name{prioritise_profiles}
\alias{prioritise_profiles}
\title{Prioritise and rank profiles}
\usage{
prioritise_profiles(
  x,
  profile_id = profile_id,
  dates = dates,
  intensity = intensity,
  max_int = 7,
  freq_det = 70,
  cum_int = 5
)
}
\arguments{
\item{x}{a tibble in long format.}

\item{profile_id}{column name of the profile IDs (default is profile_id).}

\item{dates}{column name of the dates (default is dates).}

\item{intensity}{column name of the intensity (default is intensity).}

\item{max_int}{selects profiles with at least one detection higher than x (default is 7).}

\item{freq_det}{selects profiles detected in more than x percent of the samples (default is 70).}

\item{cum_int}{selects x percent of profiles with highest cumulative intensity (default is 5).}
}
\value{
List with the following elements

* min_one_criteria: a tibble containing profile IDs, dates and intensities of
prioritised profiles that fulfill at least one criteria.

* all_criteria: a tibble containing profile IDs, dates and intensities of
prioritised profiles that fulfill all criteria.

* rank: a tibble containing profile IDs, a logical column for each criteria
indicating if it is fulfilled for the given profile and priority rank
for profiles that fulfill all criteria.
}
\description{
This function takes a dataframe with time profiles containing profile IDs (numeric or character),
dates (date format) and intensity (numeric) and returns profiles prioritised based on the following
criteria: maximum intensity, cumulative intensity, frequency of detection.
}
\details{
Priority rank: profiles that meet all three criteria are selected.
In a first step, profiles are ranked by importance for each of the three
criteria (e.g., highest maximum contribution = highest importance). The
summed-up ranks are then used to assign an overall priority rank to each
profile.
}
\references{
Chonova, T., Honti, M., Loos, M., Ruppe, S., Langlois, I.,
Griesshaber, D.,Fenner, K., Singer, H. Unveiling Industrial Contamination in
the Rhine: Insights from Data Mining of High-Frequency Measurements (in prep)
}
