% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_srcdyn.R
\name{show_friends}
\alias{show_friends}
\title{Show profiles with similar time patterns}
\usage{
show_friends(
  x,
  dist_mat,
  profile_id = profile_id,
  dates = dates,
  intensity = intensity,
  sel_prof,
  nr = 3
)
}
\arguments{
\item{x}{a tibble in long format.}

\item{dist_mat}{a distance measure (output from find_friends() or dist() function).}

\item{profile_id}{column name of the profile IDs (default is profile_id).}

\item{dates}{column name of the dates (default is dates).}

\item{intensity}{column name of the intensity (default is intensity).}

\item{sel_prof}{the ID of the selected profile.}

\item{nr}{the number of similar profiles to plot (default is 3).}
}
\value{
List with the following elements

List with the following elements

* plt: line plots of the selected time profile and profiles with time patterns similar to it.

* prof: a table with IDs of similar profiles and their respective distances to the selected profile.
}
\description{
This function takes a dataframe with profile IDs (numeric or character),
dates (date format) and intensity (numeric) as well as its distance measure,
selects profiles with similar time pattern and returns line plots.
}
